Updated according to email exchange on 9th Dec, 2022:
- rdlai2d = .true.  to use the new LAI maps from geo_em

Thanks for bringing this up. This seems to be a general Land scheme option. The dveg=8 is noah-mp specific and probably does the job, but we should probably also enable rdlaid2d just to be sure.

- dveg = 8 to use the new LAI maps from geo_em

Yes we do agree. This is what Josipa also suggested when we did the quick LAI SPOT sensitivity test.

-bl_mynn_cloudpdf was switched from 2 (default and non-gausian scheme representing non-convective clouds) to 0 (gausian scheme). Is there a reason to make this switch?

Option 0 seems a bit outdated, so we could go for the default option 2.

-cudt was switched to 5 instead of 0

We agree on 0. This is anyway the case for all cumulus schemes except KF. Playing with the cudt in KF can save time and a value around 5min is used NCAR regional climate runs. But lets go to 0 for max accuracy.

Regarding the physics options, we also suggest to use the cu_rad_feedback=.true. option. This allows sub-grid cloud fraction interaction with radiation. We have seen this can have an impact.